SAK-6294. continue... More files needed for: Default the paging number of records to 5 for audio questions. Need to take care of other types at the same time.
git-svn-id: https://source.sakaiproject.org/svn/sam/trunk@14945 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo/samigo-services/src/java/org/sakaiproject/tool/assessment/facade/PublishedAssessmentFacadeQueries.java
@@ -1512,7 +1512,7 @@ public class PublishedAssessmentFacadeQueries
 //         new org.hibernate.type.Type[] { Hibernate.LONG });
   }
 
-  public Integer getItemType(final Long publishedItemId){
+  public Long getItemType(final Long publishedItemId){
     final String query = "select p.typeId "+
                    " from PublishedItemData p "+
 	" where p.publishedItemId=?";
@@ -1530,7 +1530,7 @@ public class PublishedAssessmentFacadeQueries
 //					    new Object[] { publishedItemId },
 //					    new org.hibernate.type.Type[] { Hibernate.LONG });
     if (list.size()>0)
-	return (Integer)list.get(0);
+	return (Long)list.get(0);
     else
 	return null;
     }
