+++ samigo/samigo-services/src/java/org/sakaiproject/tool/assessment/facade/PublishedAssessmentFacadeQueriesAPI.java
@@ -284,7 +284,7 @@ public interface PublishedAssessmentFacadeQueriesAPI
 
   public List getPublishedItemIds(Long publishedAssessmentId);
 
-  public Integer getItemType(Long publishedItemId);
+  public Long getItemType(Long publishedItemId);
   
   public HashSet getSectionSetForAssessment(Long publishedAssessmentId);
